"``` (** **** Exercise: 2 stars (example_map)  *) (* Fill in the definition of example_map with a total_map that   you think example_tree should correspond to.  Use   [t_update] and [(t_empty default)]. *)
Definition example_map (v2 v4 v5: V) : total_map V := t_update (t_update (t_update (t_empty default) 2 v2) 4 v4) 5 v5.
```

This should do it.

3:55pm. Focus me. I can't believe I slipped into daydream for 20m.

https://www.jutarnji.hr/vijesti/hrvatska/analiza-otkriva-koji-je-najpozeljniji-posao-u-hrvatskoj-ne-pojavljuju-se-u-javnosti-ne-voze-skupe-automobile-ne-nose-odijela/6927975/

Ah, I was wrong in the previous entry. According to this, lead programmers (4% of them) can have salaries in excess of 2.3k. This is what I should be aiming for. Though getting to this point would require climbing a ladder, it is not realistic to get such an offer right off the bat, not with my connections and resume."