"7:50am. I got up 20m ago. Let me chill for a while and then I will start.
https://mangadex.org/title/19721/blazblue-variable-heart

Let me read this and Toji no Miko.

I still haven't finished those chapters. Vol 2 came out a few days ago.

9:15am. https://www.reddit.com/r/ProgrammingLanguages/comments/cgol52/edwin_brady_on_chez_scheme/

"Chez and SBCL are known to be fast. Given how flexible Idris' type system is it only makes sense to compile to something like Chez. He made a good choice. I've read that Idris code quality previously was not that good.

Personally, my long term wish for a dependently typed language would be to combine it with staging features. This would allow it to be compiled on restricted devices like GPUs and have very predictable performance.

I put some thought into it back in 2017 and concluded that it would be too hard, but I knew very little type theory then. I know a bit more now, but I still nowhere enough to make such an undertaking. Like with Spiral I'd want to aim for a minimalist design. Though I haven't looked under the hood, Coq and Agda feel very complex as languages. My hunch is that even if I had real expertise in dependently typed programming it would still take a few years of design work to figure it out. As simple as Spiral turned out in the end, it took me an insane amount of time and this can only be worse.

Language design is a hard thing, but I am guessing that the motivation for doing that kind of work will come from the demands of implementing more sophisticated algorithms and novel hardware. Formalizing math is also one of the motivations and the type theory community has done well following that path.

I really wish ML algorithms were better. Backprop was fun for a while, but it is old now."

Did this post.

I kind of feel like watching the talk by Brady now. Let me do that and then I will start work on the problem.

9:25am. https://youtu.be/DRq2NgeFcO0?t=503

"Guaranteed these are erased." This is actually pretty cool.

10am. Let me finally start. I got up at 7:30am and completely wasted my morning. I need to put in some time into this. From now till 6pm. I will do what I have to do."