"8:20am. I went to bed quite late yesterday, but I slept quite well so I am not tired. When I first tried Homm 4 I said that I was put off, but I am having a lot of fun with it now that I've given it a chance.
https://www.youtube.com/watch?v=lJc0Q2kJft0
Modern Physics Formalized in Modal Homotopy Type Theory by Urs Schreiber

In that thread somebody mentioned modal type theory in the context of staging and googling it I got this. Since formalizing math is my main interest at the moment, I am definitely going to watch this right now.

Well, by now, I mean as soon as I stop chilling.

8:40am. Let me watch that thing and then I will finish the chapter.

8:50am. https://youtu.be/lJc0Q2kJft0?t=405
"I'll hope to at least indicate to you for instance how fundamental theorem of calculus follows naturally from a synthetic axiom where you don't go and decompose the real line into Cauchy reals and set up differential geometry and everything, but is sort of just there."

Yeah, this is exactly what I am interesting in. The ability to do this kind of reasoning will be exactly what I need to crack credit assignment.

10:05am. https://arxiv.org/abs/1601.05956
Higher prequantum geometry

I did not get a think of that talk, but here is the 68 page paper. I'll take a look at it.

https://www.youtube.com/watch?v=ld4YL787dAk
Towards elementary infinity-toposes - Michael Shulman

Since he keeps mentioning this guy, let me also watch this talk as well. But first let me read the paper. This stuff will act as my motivation in the years ahead.

10:40am. 54/68. No way can I understand this. I just skimmed this. It just goes to show how insufficient my understanding of math is. Literally the only way somebody of my intelligence will be able to master this is if I can build it from ground up. I am not going to even try it any other way. Some people have an affinity for abstract gibberish, but I do not.

In a sense type theory is the end of everything. Unlike Spiral's staging, type theory has a program search property that will one day be its highlight.

10:50am. Let me watch the video by Mike Schulman.

11am. https://youtu.be/ld4YL787dAk?t=509
"John Baez: ...ultimately mathematics will be founded on infinity categories, with sets appearing as a simple special case.

Me: ...it may turn out that...intensional type theory...[is] a theory of infinity groupoids...which would reduce to ordinary set theory when we restrict the attention to infinity groupoids in which all hom-spaces are empty or contractible.

Peter Lumsdaine: ...asking a hom-set to be contractible involves arbitrary high dimensions, and I am not sure what kind of language ... would let you talk about that ... while a type may have infinitely high non-trivial structure, internally you can only work with it via its finite-dimensional approximations."

11:20am. https://youtu.be/ld4YL787dAk?t=1074
"An elementary topos is a category with all the finitary properties of the category Set except
* Excluded middle: monos need not have complements
* The axiom of choice: epis need not split
* Well-pointedness: the terminal object need not be a generator"

This is in contrast to elementary inf-toposes which is the title of the talk. It is good that he gave this definition as I can just think of standard inductive types as elementary toposes.

11:55am. https://youtu.be/ld4YL787dAk?t=2410

Issues with infinity seems to recur over and over throughout math.

12:30pm. Let me finally have breakfast. Every once in a while I seem to get an urge to watch lectures, but I had enough of that. When I get back I'll resume proving stuff about balanced trees. The harsh math stuff is still some ways into the future."