"11:05am. I can't believe there is still thunder and rain pouring outside. It started a few hours ago and still hasn't abated. I just got up to watch the Symphogear ep and have breakfast, but I think I'll head back to bed.
I'll start once the thunder ends.

12:55pm. It still not has receded completely, but hopefully what is left won't kill me. I need to chill for a while before starting.

...I think the thing is coming back again. ...I really won't get any work today done at this rate.

1:15pm. Well, no it is not coming back. I've almost recharged. Let me chill for a bit longer and then I will start. I will take it easy today, just studying the binomial heap code. I'll leave proofs for tomorrow.

1:30pm. I had some time to think about binomial heaps during the night. I actually went to bed on time yesterday - also in part due to bad weather, but to think it would just last for so long. At least it blew the heat wave away. That I very much welcome.

Let me put in a few words about Symphogear. I do not remember exactly how the previous seasons were, but I like how fast paced this season is. Each of the past 3 eps had two battles each, whereas shows of this type usually have only one. This is the first time I am witnessing this. The cliffhanger in the last ep was pretty good too, and I have no idea where the story will go at this rate. I did fall for the rumor that ep 4 would be grimdark and was disappointed about that, but well, I'll bear with it.

I'll save Lord Melloi for later.

1:35pm. Let me start this thing. As is usual when I am confused I will just review until I clear it up."