"9:55am. https://www.youtube.com/watch?v=cD-boHkcJO0
Let me watch this and then hopefully I will wake up in time to take a break.

10am. https://youtu.be/cD-boHkcJO0?t=333

This seems like it might be promising. I just finished learning Hoare logic a few weeks ago and here he is telling me it is no good.

10:25am. https://youtu.be/cD-boHkcJO0?t=1108

Let me stop here. I just do not feel like it right now. I need to properly chill for a while. It is this late and I still haven't properly woken up. All that energy that I had yesterday is gone.

I resent myself, I really do. I wish I was making some real progress on ML, but my ideals and reality seem to further than ever. All I really want in life is power, but somehow I am getting mired in programming.

I had a sense that it would turn out like this if I attempted it back in high school which is why I haven't done it, but now I have a heavy burden to bear.

11:35am. Done with breakfast. Let me resume the lecture. After that I will start the Lean book.

Right now, even if I wanted to do anything ML related, I really cannot do anything until I learn how real numbers are constructed in these proof assistants. Whether it be formalizing RL algorithms or anything else, without the real numbers I can't do anything. Every path leads to the reals.

So I am going to have to study a bit more before I can do anything on my own.

I guess the stressful night did its thing. I cannot bear it anymore. I want to do work in an area that interests me. Not these functional algorithms, as much as I need the exercise. Exercise I can always get. I have enough to get my foot in the door.

11:50am. No, I had enough. I'll skip the lecture as well. I just do not care about it and I am sick of Coq at this point.

Let me move to the Lean book.

I really wish I could just go to bed and get some proper rest to get rid of all my irritation and impatience.

12:25pm. I am so unmotivated right now. I've been reading manga rather than the book.

```
Alice, Alice’s husband, their son, their daughter, and Alice’s brother were involved in a murder. One of the five killed one of the other four. The following facts refer to the five people mentioned:

1. A man and a woman were together in a bar at the time of the murder.
2. The victim and the killer were together on a beach at the time of the murder.
3. One of Alice’s two children was alone at the time of the murder.
4. Alice and her husband were not together at the time of the murder.
5. The victim’s twin was not the killer.
6. The killer was younger than the victim.

Which one of the five was the victim?
```

Let me give tihs puzzle a try.

```
place time_of_murder victim beach
place time_of_murder killer beach
time_of_murder child1 alone
Alice husband both alone
victim twin not killer
```

...Yeah, screw this. I sort of got the sense I could put this into a logical notation and simplify it, but I do not see it.

```
Alice, Alice’s husband, their son, their daughter, and Alice’s brother were involved in a murder.
```

Actually, just based on what was pasted I could not figure out that there were 5 people.

Let me start with with the last point.

`6. The killer was younger than the victim.`

That immediately rules out Alice, Alice’s husband and Alice’s brother as the killers. Meaning one of Alice's children must be the killers.

`5. The victim’s twin was not the killer.`

Who could be the twins? Either Alice and her brother, or her children. This gives a lot of information. It says that the victim had a sibling and that its twin was not the killer.

So in combination...

Alice or her brother must be the victims...

No, in this chain of reasoning I am simply ignoring too much.

```
Twins?
| Alice + Alice's brother
| Alice's children
| Alice + Alice's brother & Alice's children
```

Also I've been ignoring the possibility that Alice's brother might be younger than her children in the case that they are not twins.

Let me go at it from the top.

```
1. A man and a woman were together in a bar at the time of the murder.
2. The victim and the killer were together on a beach at the time of the murder.
3. One of Alice’s two children was alone at the time of the murder.
4. Alice and her husband were not together at the time of the murder.
5. The victim’s twin was not the killer.
6. The killer was younger than the victim.
```

Let me think about the time of murders for a bit here.

Since Alice and her husband were not together at the time of the murder that narrows down the man and the woman who were in a bar.

```
Man and woman in a bar at the time of murder?
~Alice + Alice's husband~
| Alice + Alice' brother
| One of Alice's children + Alice's brother
```

In the second case, I would be able to infer that one of Alice's children is a woman then. But let me put more of the facts together.

```
3. One of Alice’s two children was alone at the time of the murder.
```

I will assume that this means that one of her children cannot be either the victim or the killer. That means I can completely leave it out of my reasoning.

```
2. The victim and the killer were together on a beach at the time of the murder.
```

Who could the victim and the killer be?

```
Man and woman in a bar at the time of murder?
| Alice + Alice' brother -> One of Alice's children + Alice's husband
| One of Alice's children + Alice's brother -> ???
```

In the second case I cannot think of anything that might fit. Alice's other child cannot be there as it was alone. Alice and her husband were also not together at the time of murder.

Therefore...

```
One of Alice's children + Alice's husband
```

Is the correct choice.

```
5. The victim’s twin was not the killer.
6. The killer was younger than the victim.
```

Based on this, the father is the victim and the child is the culprit.

1pm. Naruhodo. I guess it was a fun little puzzle.

It uplifted my spirit a little. I guess I'll dive into the chapter properly now."