"10:10am. I really got up late today, but boy did I sleep well. A complete difference from yesterday.
https://mangadex.org/title/9828/aiki-s

I forgot about this one. It got updated after two years.

10:15am. Well, the plan this morning is to chill a while and then I'll check out the papers. When I am done with them I'll start looking into Cubical Agda. Presumably since it exists, they have some examples of HITs in action. I'll have to study them.

10:45am. Let me start on those papers.

https://jozefg.github.io/papers/2019-implementing-modal-dependent-type-theory.pdf

Let me start with the first which is only 30 pages. The other is 90.

I won't spend too much time on this. I just want to get a sense of this.

11:05am. 9/29. I am going to have to look at some real examples on how this works. This is just not enough for me to really understand what this is about.

11:10am. The general feel of this does evoke similarity to staging.

11:30am. "Applications in mathematics. A current goal within the scientific community is to maximize
the amount of mathematics which can be formalized synthetically inside type theory [Shulman
2018; Univalent Foundations Program 2013], avoiding low-level analytic details. A recent success
story involving the necessity modality in dependent type theory can be found in the internal
construction of classifying objects for fibrations in models of homotopy type theory based on a
“tiny intervalž [Angiuli et al. 2019; Licata et al. 2018]. This construction uses a right adjoint to the
path space functor (−)I
in a critical way, but this functor cannot be internalized. Using the necessity
modality of MLTTµ, it is easy to axiomatize this adjunction and use it to construct the classifying
fibration. Licata et al. [2018] have presented a formalization in agda-flat, and we conjecture that it
should be possible to formalize the same in MLTTµ. We hope that such a formalization may benefit
from the additional definitional equalities enabled by our treatment of the modality.
More generally, the necessity modality of MLTTµ enables the type-theoretic axiomatization of
global operations (such as comonadic modalities) which are not closed under substitution. This
technique makes it possible to formalize a great deal of mathematics inside type theory, and the
implementability of MLTTµ promises the benefits of machine-checking for these formalizations."

Well, I still do not know what these modalities are good for, but I guess I'll take their word for it.

https://arxiv.org/abs/1706.07526
Modalities in homotopy type theory

Let me check this paper out next. Though after I am done with it, there will be no more dallying. I'll dive into Cubical Agda properly.

11:40am. No, this paper is too much. It is just an unending stream of lemmas and theorems.

11:50am. Let me stop here so I can have breakfast.

https://www.oreilly.com/ideas/why-a-data-scientist-is-not-a-data-engineer
https://www.jesse-anderson.com/2017/03/what-happens-when-you-hire-a-data-scientist-without-a-data-engineer/

I guess these articles should serve as reading while that is going on."