"9:10am. In contrast to yesterday, I really slept quite well during the night. Let me do some chilling and I will start. I am going to have to point the author at the SO question. I'll do it in the thread I've opened and then closed. I'll tell him about the typo as well.
Yesterday my focus was quite low, but after stressing out about it for a while I can feel my motivation rising. It is always like this. I can never just flip and enter a completely different field right away. There is always some period of adjustment.

9:25am. Let me start this thing.

9:45am. Passed along the question. For the time being let me focus on the second half.

```
The first thing Cubical Agda add is an interval type I. Then, we add the PathP types that behave
like function types out of the interval, but with fixed endpoints. Note that the interval in Cubical
Agda is not inductively defined, so we cannot pattern match on it. This follows from the intuition
that path types are continuous functions from the interval into a space so that they cannot provide
arbitrarily different results for i0 and i1.
```

Huh really. This is interesting. I was sure it was inductively defined based on how it was used.

10:10am. 15/29. I'll paste this stuff into the file later...No actually let me do it now. I'll do everything in turn.

10:40am. I keep wandering into daydreams, but...

`transport (λ j → B j (transp (λ i → A (j ∨ ~ i)) j x)) (f (transport (λ i → A (~ i)) x))`

I get most of this. I do not get `transp (λ i → A (j ∨ ~ i)) j x` in particular, but I guess I'll leave that for later. I do not get why passing i0 and i1 to `transp` is not symmetric.

12:40pm. Done with breakfast. My focus has been low as I've been in mourning. It is not the same feeling as yesterday, I do feel a bit more optimistic today, but I still need more time to deal with it. I could cope with it while I was doing the exercises, but going from the kind of mindset that I had while working on Spiral.

The exercises while tough were fun and games. But what I will be doing here is abandoning my old ways. In a sense, it will be like packing up and moving to a new continent. I cannot help but feel a great sense of loss that touches upon all my past failures and regrets.

I loved my old way of thinking. Spiral was pretty much a love song to that.

I cannot help, but feel a certain sense of betrayal by the ML community for letting it come to this. I really have convinced myself that I should have won with Spiral.

If I have to make yet another large step forward then...

I guess after all this time, I need to change one more time. One more time I will change, and this time so will the nature of my talent.

1:15pm. Just like yesterday, I think I will chill for a while. Maybe I'll listen to music, or maybe I'll step away from the screen for a while. During the past two days, I spent most of my working day aimlessly lurking 4chan.

The one thing I won't do, at least till 6pm is play games. Games are too engrossing and would trivialize this whole affair. I need to to work out my internals. Outwardly and logically, it does not really make sense to be depressed over studying cubical type theory, but I do not really understand the whys of my inner processes.

I'll slack for a while and then once I am really ready, I will go at it full force. This might be tomorrow, or this might be a few days from now. Either way, boredom will sink in eventually and force me to move."