"3:55pm. 22/29. Due to the programming session my focus is really low. This stuff is definitely over my head and would be even if I was at full power.
4:15pm. I could almost call it a day here. I am not even close to in the mood for study, but I can't really stop just yet.

I guess I'll study the library then. There should be simple examples out there.

4:30pm. Ahhh, I see there is a BinNat in the library! I wish that guy pointed it out to me, but nevermind that.

'There is in fact a `Cubical.Data.BinNat` module in the Cubical library that has all the proofs for binary conversion in cubical style. This answers my question.'

I added this comment to the question. And I've closed the issue I've opened.

4:45pm. Definitely can't study this right now. I am too tired. Tomorrow, studying `BinNat` will be my main topic. I'll translate it into my own style as exercise.

Until I understand what everything in the primitives section does and how it can be used, I am going to keep going at it. Now that I've found this, I have reason to think that everything else in the paper will be in the library as well.

4:50pm. I've had to do soul searching over the past few days, but I think that having the goal of studying the Cubical library as a part of mastering CubeTT is just fine. I do not need anything more or anything less.

In a month or two I'll be done and I'll move to the next thing, but the present and the near term are the time for this to be done. I have no need for shame or doubt.

4:55pm. Now let me have fun. I've been hearing rumbling of the thunder in the distance, but maybe it will leave me alone for the rest of the day."