"8:15am. Got up 15m ago and am chilling right now. I'll start in a while.
So Granbelm is pretty good as it turns out. Anime about mages being good seems to be a theme for this season together with Lord Melloi. For the later, now I've realized that the manga has over 10 chapters out and am reading it.

It feels like things are looking up for me now that I've figured out the cubical binary conversion. Today I'll advance my knowledge even further. Recently I've given some though to dependent language design. It definitely is a challenge to figure out how to merge a top-down dependent type system with a a bottom-up staging one, but it does not feel like an impossible task.

I've made up my mind - I won't do it now as I have no need of it, but if type theory can give me the algorithm that I seek, I will repay my gratitude by making Spiral v0.2 with dependent types. It would be a win-win for everybody. Well, apart from the online poker pro community, but who care about them.

My motivation for making v0.1 was for the sake of ML, so now that I do not have anything to use it with, I definitely have no motivation to take on such a mission.

Damn it, there is really so much that I do not understand. Looking past the Singularity, if AI is about optimizing simple models, then what does that leave for the post-humans? If the grand theory of learning is attainable to peons such as myself, what will that leave for the Inspired me?

On one hand, I really do hope that the path towards Singularity will not take centuries of breakthroughs, but on the other hand it being too easy on the tier of deep learning will not leave anything either.

What I really, really want is to ditch my current way of programming where I am always handcrafting everything and replace it with something more generic where I am essentially manipulating and composing models of learning.

Reconciling those two desire leads me to hope that the theory of learning will proceed is through the study of model families and their optimization procedures. Neural nets + backprop are an instance of it. Fundamental to this vision is that an optimization procedure should be dependent on the structure and the properties of a particular model.

The reason why backprop is bad is because it is blind too that structure, much like humans at this point.

8:35am. I need to find a new, higher way of thinking and I need help to get to that point. To maximize my chances I need to take advantage of computers, polish my fundamentals and then with the aid of that gain a proper understanding of all the fields of mathematics that have been discovered since the 17th century. I need to advance beyond that time. Newton will not cut it.

8:40am. It is simply a matter of principle of what I have to do next. Simply thinking about it won't get me any closer to the answer here. There is so knowledge out there and the only way to surmount the block is to integrate it into proper understanding.

I think that ultimately, learning will turn out to be much like physics and as we dig deeper there will be algorithmic links between the learning systems we will create and the physical reality. This will fuel the coming awareness that the universe is some kind of learning system.

The principles of learning starting with evolution are ubiquitous, so it is that much more annoying to get hung up at this point.

8:45am. There is no real end to inspiration to draw from if one looks around. Maybe a really pertinent question is why learning has not been cracked in spite of that?

My guess is - the fundamentals were insufficient. It is a long path from dirt to DNA. Isn't it great then that HoTT got discovered just in the past several years?

8:50am. I need to graduate from being the ultimate code monkey to something more. Yeah, I was good at programming, but that means nothing and I knew it. Mathematics was always a blade hanging over my neck. I never ever did it properly. I never trusted it. I was never comfortable with it. For good reasons. And consequently I was bad at it. It was a negative feedback loop.

I trusted code, but code was far, far too weak for me to take it seriously.

I've made up my mind. Even if close the gap with the top ML researchers and stand at the pinnacle of both programming and ML, that might not give me the power that I truly desire. In this era, even with the best skills in both worlds, I might still remain as the ultimate code monkey rather than become the manipulator of learning for the rest of my life.

8:55am. But I think that maybe as a goal that might not be too bad. Not that many humans would be able to attain such a level of skill. If I am destined to fail let it be unfair, over things outside my control.

I am hopeless. I despise this world for not leaving me any paths that are even remotely connected to normality. Money? Pfffttt. It would get in the way of learning. Social life? Has nothing to do with programming. Job? Only if I am desperate.

If I crack this, normies will have only themselves to blame for suddenly finding themselves the object of replacement. Today's world is proof that false hope can push out the true hope. That injustice needs to be corrected.

9:05am. Yeah, I get it. I though I was being crazy when I wrote it, but in fact the goal of exceeding Deepmind is the one that is attainable. ML research is in fact the one that is out of reach.

9:10am. I really should have been reading manga for the past 45m instead of ranting so let me get back to it. I'll resume studying the library after that."