"9:45am. Let me stop here. I want to start the studies.
9:55am. Focus me, focus. Stop reading the Fate thread on /a/.

10am.

```
-- We can then also transport the property
oddℕnotEvenℕ : (n : ℕ) → oddℕ n ≡ not (evenℕ n)
oddℕnotEvenℕ =
  let -- We first build a path from oddBinℕ to oddℕ. When i=1 this is
      -- "transp (λ j → Binℕ≡ℕ j → Bool) i0 oddBinℕ" (i.e. oddℕ)
      oddp : PathP (λ i → Binℕ≡ℕ i → Bool) oddBinℕ oddℕ
      oddp i = transp (λ j → Binℕ≡ℕ (i ∧ j) → Bool) (~ i) oddBinℕ

      -- We then build a path from evenBinℕ to evenℕ
      evenp : PathP (λ i → Binℕ≡ℕ i → Bool) evenBinℕ evenℕ
      evenp i = transp (λ j → Binℕ≡ℕ (i ∧ j) → Bool) (~ i) evenBinℕ
  in -- Then transport oddBinℕnotEvenBinℕ in a suitable equality type
     -- When i=0 this is "(n : Binℕ) → oddBinℕ n ≡ not (evenBinℕ n)"
     -- When i=1 this is "(n : ℕ) → oddℕ n ≡ not (evenℕ n)"
     transport (λ i → (n : Binℕ≡ℕ i) → oddp i n ≡ not (evenp i n)) oddBinℕnotEvenBinℕ
```

I need to focus on this for a while.

10:10am.

```
  let -- We first build a path from oddBinℕ to oddℕ. When i=1 this is
      -- "transp (λ j → Binℕ≡ℕ j → Bool) i0 oddBinℕ" (i.e. oddℕ)
      oddp : PathP (λ i → Binℕ≡ℕ i → Bool) oddBinℕ oddℕ
      oddp i = transp (λ j → Binℕ≡ℕ (i ∧ j) → Bool) (~ i) oddBinℕ
```

I need to iterate the 4 cases of `i`. I do not really understand what is going on here otherwise.

```
i = i0
j = i0
oddp : PathP (Binℕ → Bool) oddBinℕ oddℕ
oddp = transp (Binℕ → Bool) i1 oddBinℕ
```

I think the paper mentioned that when the second argument to `transp` is `i1` the function acts like the identity.

I am going to have to experiment a bit with `transp`.

```
{-# OPTIONS --cubical #-}

open import Cubical.Core.Everything

transport : ∀ {A B : Set} → A ≡ B → A → B
transport p a = transp (λ i → p i) i0 a
```

I remember playing with this before. If it is really the identity then...

```
transport : ∀ {A B : Set} → A ≡ B → A → A
transport p a = transp (λ i → p i) i1 a
```

No, this gives me a type error.

```
This is defined using another primitive of Cubical Agda called transp. It is a generalization of the
regular transport principle which lets us specify where the transport is the identity function. In
particular, when the second argument to transp is i1 it will reduce to a, which let us prove that
transp A r a is always path equal to a (cf. addp later).
```

```
transport : ∀ {A B : Set} → A ≡ B → A → A
transport p a = transp (λ i → p i0) i1 a
```

Yeah, this passes, but in this case it does not make any difference whether the second argument is `i1` or `i0`.

What is this `addp`?

```
addp : PathP (λ i → ℕ≡Bin i → ℕ≡Bin i → ℕ≡Bin i) _+_ _+Bin_
addp i = transp (λ j → ℕ≡Bin (i ∧ j) → ℕ≡Bin (i ∧ j) → ℕ≡Bin (i ∧ j)) (~ i) _+_
```

No, I do not get this. How about I ask?

10:40am. I made a question. I guess I'll get a reply at some point. Let me read the Kaiji chapter.

10:50am. I guess I'll have breakfast here. Until I figure this thing out I have no choice but to keep playing around with this."