"12:45pm. Done with chores.
Let me read Ribbon Warrior and then I'll resume.

1:50pm.

```
_≃'_ : ∀ {ℓ} (A B : Set ℓ) → Set ℓ
A ≃' B = Σ[ f ∈ (A → B) ] (∀ (y : B) → Σ[ p ∈ (Σ[ x ∈ A ] f x ≡ y) ] (∀ p' → p ≡ p'))
```

I've thought about this for a while. I think I understand it completely now.

Let me first talk about `∀ p' → p ≡ p'`. At first glance it seems confusing that this would represent a one to one mapping.

First let me highlight that `p` and `p'` are tuples `x, y` where `x` represents the start point and y the end point. More precisely they dependent products of type `Σ[ x ∈ A ] f x ≡ y`, but for the purposes of explanation it might be better to ignore that and to consider `p ≡ p'` as `(x,y) ≡ (x',y')`. So in order for this equality to hold that means that `x ≡ x'`  and `y ≡ y'` needs to hold.

Let me consider the right sides of `p` and `p'` first, meaning I am talking about `y` and `y'`.

Suppose `y` is `3`. That means that `y'` would need to be 3 as well. This is straightforward and is how one would expect that equality would behave.

Now let me consider the left side, the `x` and `x'`. Obviously they need to be propositionally equal, but here is where a certain subtlety comes into play, namely the `∀` in `∀ p' → p ≡ p'`. This is actually the part that ensures that this is a one to one mapping.

It is easy to see that this holds for the identity, but let me assume that `f` is something like `fun x => x / 2`. That would mean that `0 => 0` and `1 => 0`. Meaning a proof would be needed that both `(0,0) ≡ (0,0)` and `(0,0) ≡ (1,0)`. That `∀ p'` could be interpreted as enumerating the starting point `x` for each `y`.

2:10pm. Informally what equivalence means is that given `f : A -> B`  that for all `y : B` that `f x ≡ y` holds for each starting point `x` and that the starting points all be unique.

2:15pm. Let me write out the definition once without looking...

`A equiv B := forall (f : A -> B), exists (p : (exists (x : A), f x = y)), forall p', p = p'`

I am sure of everything but that `forall f`...

`A equiv B := exists (f : A -> B), forall (y : B), exists (p : (exists (x : A), f x = y)), forall p', p = p'`.

Here it is. I forgot the `y`, but that is minor. Mistaking `forall f` and `exists f` is what is inexcusable. The first one would mean, give me an `f` and I will give you everything else. No way could the be constructed.

The irony is that reasoning about the contents of the body in `exists f, ...` and `forall f, ...` is much the same as once you are inside as `f` is some specific instance and it is good to forget whether or not it is supposed to be a loop. Loop reasoning comes afterwards.

My issue was that I forgot that I am writing out a type, the equivalence definition is not an instance of an equivalence, but its type.

2:30pm. I've prettied up the attempt to look more like Coq pseudo-code.

2:35pm. This is great. I am doing some final passes in my head before I can properly internalize this.

`A equiv B := exists (f : A -> B), forall (y : B), forall (p : (exists (x : A), f x = y)), forall p', p = p'`.

Going by my logic would this definition not be fine as well? Let me think about it while I take a bath."