"3:20pm. Had some time to think about it. Well, it is simple.
Suppose is `y = 3` and I have to prove that `(a,3) = (a, 3)` and `(a,3) = (b,3)`. If I have proof that `a = b` lying around then this is doable.

If the definition is the one with the existential - ``A equiv B := exists (f : A -> B), forall (y : B), exists (p : (exists (x : A), f x = y)), forall p', p = p'`` then I just have to prove `(a,3) = (a, 3)` and `(a,3) = (b,3)` and I'll be done with it.

The forall version would require me to prove `(a,3) = (a, 3)` and `(a,3) = (b,3)` and `(b,3) = (a, 3)` and `(b,3) = (b,3)`. I would not be getting any more coverage, just more work. The existential version is definitely more efficient.

3:25pm. The definition is definitely the right one. It is definitely one for the books. I am going to have to do more study in order to figure out why isomorphism is only a special case of equivalence and I'll get to that in due time.

3:30pm. Now, I need to study that identity equivalence proof.

```
((y , refl) , λ z i → z .snd (~ i) , λ j → z .snd (~ i ∨ j))
```

Let me try doing this without looking at the solution.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y =
  let p = (y , refl) in p , λ { (y' , refl') → {!refl'!}}
```

```
Goal: (y , (λ _ → idfun A y)) ≡ (y' , refl')
————————————————————————————————————————————————————————————
p     : Σ A (λ x → idfun A x ≡ y)
refl' : idfun A y' ≡ y
y'    : A
y     : A
A     : Set ℓ
ℓ     : Level  (not in scope)
```

Hmmm...

I think this is reasonable. Let me give it a try.

3:45pm.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y =
  let p = (y , refl) in p , λ { (y' , refl') i → refl' (~ i) , λ j → {!refl' (~ i ∧ j)!}}
```

```
Goal: A
————————————————————————————————————————————————————————————
p     : Σ A (λ x → idfun A x ≡ y)
j     : I
i     : I
refl' : idfun A y' ≡ y
y'    : A
y     : A
A     : Set ℓ
ℓ     : Level  (not in scope)
———— Constraints ———————————————————————————————————————————
[30] (?0 (j = i1)) = y : A
[30] (?0 (j = i0)) = (refl' (~ i)) : A
```

I find this part really confusing.

Let me backtrack.

3:50pm.

```
Goal: A
————————————————————————————————————————————————————————————
p     : Σ A (λ x → idfun A x ≡ y)
i     : I
refl' : idfun A y' ≡ y
y'    : A
y     : A
A     : Set ℓ
ℓ     : Level  (not in scope)
———— Constraints ———————————————————————————————————————————
[29]
  y = (?0 (A = A) (y = y) (y' = fst₁) (refl' = snd₁) (i = i0)) : A
[27]
  fst₁ = (?0 (A = A) (y = y) (y' = fst₁) (refl' = snd₁) (i = i1)) : A
```

What exactly am I trying to prove here again? The constraints are not using the right names.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y =
  (y , refl) , λ { (fst₁ , snd₁) i → (snd₁ (~ i)) , λ i₁ → {!snd₁ (i₁ ∨ )!} }
```

```
Goal: A
————————————————————————————————————————————————————————————
i₁   : I
i    : I
snd₁ : idfun A fst₁ ≡ y
fst₁ : A
y    : A
A    : Set ℓ
ℓ    : Level  (not in scope)
———— Constraints ———————————————————————————————————————————
[25] (?0 (i₁ = i1)) = y : A
[25] (?0 (i₁ = i0)) = (snd₁ (~ i)) : A
```

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y =
  (y , refl) , λ { (fst₁ , snd₁) i → (snd₁ (~ i)) , λ i₁ → snd₁ (i₁ ∨ ~ i) }
```

The reasoning for this is straightforward. When `i₁ = i1` the left side of the or does its thing. And when `i₁ = i0` the right side of the or satisfies the constraint.

I sort of remembered it as being and, but when I thought about it, I managed to get the answer on my own.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y = (y , refl) , λ { (y' , refl') i → (refl' (~ i)) , λ j → {!!} }
```
```
Goal: A
————————————————————————————————————————————————————————————
j     : I
i     : I
refl' : idfun A y' ≡ y
y'    : A
y     : A
A     : Set ℓ
ℓ     : Level  (not in scope)
———— Constraints ———————————————————————————————————————————
[25] (?0 (j = i1)) = y : A
[25] (?0 (j = i0)) = (refl' (~ i)) : A
```

I was going to open an issue, but now it is printing things properly it seems.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y = (y , refl) , λ { (y' , refl') i → (refl' (~ i)) , λ j → refl' (j ∨ ~ i) }
```

Let me finally compare it with the solution as given.

```
idIsEquiv : ∀ {ℓ} (A : Set ℓ) → isEquiv (idfun A)
equiv-proof (idIsEquiv A) y = ((y , refl) , λ z i → z .snd (~ i) , λ j → z .snd (~ i ∨ j))
```

4:25pm. The order given in the or does not seem to matter. Well that is what one would expect, but putting j first does make it easier on my brain.

Ok...

Also I need to keep in mind that tuples are records in Agda for future reference.

4:30pm. Hmmm...I am thinking. Got a bunch of replies on that interval question. So the point is that `u0` and `u1` are supposed to be used to represent the same thing? That is what I got from the reply.

4:45pm.

```
ua : ∀ {ℓ} {A B : Set ℓ} → A ≃ B → A ≡ B
ua {_} {A} {B} e i = Glue B (λ { (i = i0) → (A , e)
                               ; (i = i1) → (B , idEquiv B) })
```
```
uaβ : ∀ {ℓ} {A B : Set ℓ} (e : A ≃ B) (x : A) → transport (ua e) x ≡ e .fst x
uaβ e x = transportRefl (e .fst x)
```

I actually do understand this.

5:05pm. I am almost ready to call it a day. I'll go get lunch soon.

At this point I think I've gotten all I can out of the documentation. I think that right now all I am missing is an understanding of homogenous composition and the stuff related to it.

If I can make this click I can make everything else go through as well. Once I get this groundwork out of the way, the higher inductive types won't give me too much trouble.

At that point I will be able to start figuring out why `S1 = Z` and so on.

What I will do is go back to the paper I was reading before. I got some exp now so it should be easier to follow.

https://arxiv.org/abs/1611.02108
Cubical Type Theory - a constructive interpretation of the univalence axiom

I'll add this to my reading list as well. No doubt there should be something in here to help me understand all the stuff that is in Agda currently.

5:10pm. I think I will take off here.

Just a few more steps and I will get through this very basic stuff and will be able to start studying the stuff in the library properly. I am really hoping that this effort will pay off down the road when I start studying ML again."